Warm-start fnnls so an all-positive scene costs one solve
Lawson-Hanson begins at x = 0 and admits one component per iteration. A scene
whose solution is entirely positive -- the common case -- therefore paid n
growing dense solves to arrive at what a single solve gives: 4.0 s against
0.077 s for spsolve at 1000 templates, and this is the default fit method now.

`fnnls` tries the unconstrained answer first and keeps it when it satisfies the
KKT conditions. That is a proof of optimality, not a heuristic, so the shortcut
cannot return a different solution from the long route -- the same test decides
both. An all-positive scene of 1000 templates now takes 0.012 s, faster than
spsolve, because a dense solve on a well-conditioned system beats the sparse
factorization at that size.

A bounded refinement follows for partly pinned scenes: drop whatever the solve
wanted negative, re-solve, re-test, up to ten rounds. Every candidate goes
through the same KKT check, so a guess that does not survive is discarded and
the exact loop runs from scratch.

Measured on realistic scene coupling, where a template overlaps only its
neighbours rather than the whole scene: the worst case is 0.074 s at 500
templates with 30% pinned, and 0.024 s at 1000. The large ratios in a dense
random benchmark come from every template overlapping every other, which no
real scene does.

Three tests added: the warm start agrees with scipy.optimize.nnls where many
components are pinned, an all-positive scene of 300 templates returns the plain
solve, and the joint system never pins a shift while still letting one go
negative.
