+++ STATUS.md
@@ -3,6 +3,48 @@
 This file records completed implementations, validation runs, and the current work state.
 
 ## Current Work
+- [x] Shrank the test suite to the contracts a future change can break
+  (2026-08-16). 41 files / 14,088 lines / 72 s became **33 files / 11,407
+  lines / 46 s**, and the suite is green: 471 passed, 0 failed. Nothing that
+  pins a behaviour was dropped without its coverage landing somewhere else.
+
+  Deleted outright: `test_scene_astrometry.py` (an empty file), and
+  `test_sed_estimator_experiment.py`, whose module lives in the gitignored
+  `scratch/` tree so it could only ever skip. `test_sed_stack.py` went too --
+  `sed_stack.py` is a leaf nothing in the package imports.
+
+  Merged into the module that owns them: `test_fit.py` into
+  `test_scene_fitter.py`; `test_pipeline_dedup.py` and
+  `test_pipeline_multitemplate.py` into `test_pipeline.py`;
+  `test_background_masking.py` and the old `test_catalog.py` into one
+  `test_catalog.py`. Cross-file duplicates went with them: `get_bg_and_ivar`'s
+  masking was asserted in three files and is now in `test_catalog.py` alone,
+  and `as_label_array` was tested weakly in `test_templates.py` and
+  thoroughly in `test_memory_footprint.py` -- the thorough four moved to
+  `test_templates.py` and the weak one is gone.
+
+  `test_moffat_recovery.py` became `examples/verify_moffat_recovery.py`. It
+  was 24 s of the suite -- a third of it -- to produce diagnostic PNGs and
+  assert only that the median flux ratio is somewhere near unity. It is a
+  validation run, not a unit test, and now runs on demand
+  (`python examples/verify_moffat_recovery.py [outdir]`); verified end to end,
+  all five scenarios, median ratio 0.9998-1.0007.
+
+  The astrometry files were the exception to the plan. Four files, 2,199
+  lines, were the obvious consolidation target, but `test_astrom_robust.py`
+  (leaf statistics), `test_scene_astrometry_blocks.py` (blocks against a dense
+  design) and `test_scene_astrometry_robust.py` (per-anchor measurement) are
+  three layers of one argument rather than three copies of it -- collapsing
+  them would have lost the layer that localizes a failure. The redundancy was
+  in `test_astrometry.py` instead: four end-to-end scene tests asserting a
+  0.6 px shift to `atol=0.3`, where the scene files assert the same properties
+  to 1e-9. Those four went, `test_astrometry.py` is now the non-joint
+  `AstroCorrect`/`AstroMap` path plus the alpha0 scaling nothing else covers,
+  and its write into `Path("../tmp")` -- outside the repository -- went with
+  them.
+
+  Left alone deliberately: `tests/test_nnls.py`, live work on this branch.
+
 - [x] `fit_method` replaces `positivity`, and NNLS is the default (2026-08-16).
   One switch with three values: `"lls"` keeps negative fluxes, `"clip"` clamps
   them after an unconstrained solve, `"nnls"` solves under the constraint.
@@ -39,6 +81,22 @@ This file records completed implementations, validation runs, and the current wo
   `clip` and `nnls` agree on this weakly blended mock and diverge in strong
   blends, which is what `tests/test_nnls.py` covers with an 85%-correlated
   pair.
+
+  Lawson-Hanson starts at `x = 0` and admits one component per iteration, so a
+  scene whose solution is entirely positive -- the common case -- would pay `n`
+  growing solves to reach what one solve gives: 4.0 s against 0.077 s for
+  `spsolve` at 1000 templates. `fnnls` now tries the unconstrained answer
+  first and keeps it when it satisfies the KKT conditions, which is a proof of
+  optimality rather than a guess, so the shortcut cannot return a different
+  solution from the long route. That case is now 0.012 s, faster than
+  `spsolve`. A bounded refinement (drop what came out negative, re-solve,
+  re-test) handles partly pinned scenes, and anything that fails the test falls
+  back to the exact loop.
+
+  On realistic scene coupling -- a template overlapping only its neighbours --
+  the worst measured case is 0.074 s at 500 templates with 30% pinned. The
+  alarming ratios in a dense random benchmark come from every template
+  overlapping every other, which no real scene does.
 
   `info["at_bound"]` marks components sitting on the bound. Their `err` is the
   unconstrained `sqrt(diag(A^-1))`, which is not a symmetric 1-sigma interval
