+++ tests/test_astrometry.py
@@ -1,3 +1,14 @@
+"""The non-joint astrometry path: ``AstroCorrect``, ``AstroMap`` and the
+Chebyshev basis they share.
+
+The joint scene solver is covered exactly, and to much tighter tolerances,
+in ``test_scene_astrometry_blocks.py`` (the normal-equation blocks against a
+dense design) and ``test_scene_astrometry_robust.py`` (per-anchor measurement
+and robust weighting). What is left here is the separate-step estimator that
+those two never touch, plus the one scaling property of the shift block that
+only shows when ``alpha0`` is varied by hand.
+"""
+
 import os
 import sys
 
@@ -7,22 +18,14 @@ sys.path.append(os.path.join(os.path.dirname(__file__), "."))
 import numpy as np
 from numpy.polynomial.chebyshev import chebval
 from scipy.ndimage import shift as nd_shift
-from scipy.signal import fftconvolve
 
-from pathlib import Path
 from utils import make_simple_data
 from mophongo.psf import PSF
 from mophongo.templates import Templates, Template
 from mophongo.fit import SparseFitter, FitConfig
 from mophongo.astrometry import AstroCorrect, AstroMap, cheb_basis
-from mophongo.scene import generate_scenes, make_scene_basis, assemble_scene_system_AB
+from mophongo.scene import make_scene_basis, assemble_scene_system_AB
 from mophongo.scene_fitter import SceneFitter, build_normal
-from utils import save_diagnostic_image
-
-
-# ---------------------------------------------------------------------------
-# Shared helpers for the scene-solver astrometry tests
-# ---------------------------------------------------------------------------
 
 
 def _assign_solution(fitter, sol) -> None:
@@ -33,43 +36,12 @@ def _assign_solution(fitter, sol) -> None:
         tmpl.err = float(err)
 
 
-def _make_gaussian_psf_array(fwhm: float, size: int = 51) -> np.ndarray:
-    """Return a normalised 2-D Gaussian PSF of given FWHM (in pixels)."""
-    sig = fwhm / 2.355
-    y, x = np.mgrid[:size, :size] - size // 2
-    p = np.exp(-0.5 * (y ** 2 + x ** 2) / sig ** 2)
-    return p / p.sum()
+# ---------------------------------------------------------------------------
+# AstroCorrect: the separate-step estimator
+# ---------------------------------------------------------------------------
 
 
-def _collect_to_shifts(scenes) -> tuple[np.ndarray, np.ndarray]:
-    """Return arrays of (dx, dy) stored in tmpl.to_shift across all scenes."""
-    dx_list, dy_list = [], []
-    for scene in scenes:
-        for tmpl in scene.templates:
-            ts = getattr(tmpl, "to_shift", None)
-            if ts is not None and (abs(ts[0]) > 1e-6 or abs(ts[1]) > 1e-6):
-                dx_list.append(float(ts[0]))
-                dy_list.append(float(ts[1]))
-    return np.array(dx_list), np.array(dy_list)
-
-
-def _run_scenes(templates, science, weight, cfg):
-    """Generate and solve all scenes; return the scene list."""
-    scenes, _ = generate_scenes(
-        templates,
-        science,
-        weight,
-        coupling_thresh=cfg.scene_coupling_thresh,
-        astrom_minimum_snr=cfg.astrom_minimum_snr,
-        minimum_bright=cfg.scene_minimum_anchors,
-    )
-    for scene in scenes:
-        scene.set_band(science, weight, config=cfg)
-        scene.solve(config=cfg, apply_shifts=False)
-    return scenes
-
-
-def test_polynomial_astrometry_reduces_residual(tmp_path):
+def test_polynomial_astrometry_reduces_residual():
     images, segmap, catalog, psfs, truth, wht = make_simple_data(
         nsrc=10, size=151, peak_snr=5, seed=42
     )
@@ -95,26 +67,8 @@ def test_polynomial_astrometry_reduces_residual(tmp_path):
 
     rhx, rhy = ac(np.array([[50.0, 50.0]]))
 
-    print(f"Astrometry correction shifts: {rhx}, {rhy}")
-    print(f"input shifts: {shx}, {shy}")
-
     assert abs(rhx[0] - shx) < 0.3
     assert abs(rhy[0] - shy) < 0.3
-
-    tmp_path = Path("../tmp")
-    tmp_path.mkdir(exist_ok=True)
-    fname = tmp_path / "diagnostic_poly_shift.png"
-    model = images[1] - res0
-    save_diagnostic_image(
-        fname,
-        truth,
-        images[0],
-        images[1],
-        model,
-        res0,
-        segmap=segmap,
-        catalog=catalog,
-    )
 
 
 def test_gp_astrometry_returns_models():
@@ -157,6 +111,11 @@ def test_astromap_recovers_shift():
     dx, dy = amap(np.array([[75.0, 75.0]]))
     assert abs(dx[0] - shx) < 0.3
     assert abs(dy[0] - shy) < 0.3
+
+
+# ---------------------------------------------------------------------------
+# applying a solved field, and the basis it is expressed in
+# ---------------------------------------------------------------------------
 
 
 def test_apply_template_shifts_uses_shift_field():
@@ -210,139 +169,21 @@ def test_cheb_basis_handles_domain_edges():
 
 
 # ---------------------------------------------------------------------------
-# Test 1: end-to-end scene-solver shift recovery
+# the one shift-block property that only a hand-scaled alpha0 exposes
 # ---------------------------------------------------------------------------
 
-def test_scene_solver_recovers_known_shift():
-    """Scene.solve() recovers a known global pixel shift applied to the science image.
-
-    A global (dx, dy) shift is applied to the low-resolution science image.  After
-    running generate_scenes + Scene.solve(apply_shifts=False) the per-template
-    to_shift values should have a mean close to the injected shift.
-    """
-    images, segmap, catalog, psfs, truth, wht = make_simple_data(
-        nsrc=15, size=151, peak_snr=20, seed=42
-    )
-
-    psf_hi = PSF.from_array(psfs[0])
-    psf_lo = PSF.from_array(psfs[1])
-    kernel = psf_hi.matching_kernel(psf_lo)
-
-    true_dx, true_dy = 0.6, -0.5
-    science = nd_shift(images[1], (true_dy, true_dx))
-    weight = wht[1]
-
-    positions = list(zip(catalog["x"], catalog["y"]))
-    tmpls = Templates.from_image(images[0], segmap, positions, kernel)
-
-    cfg = FitConfig(
-        fit_astrometry_joint=True,
-        astrom_minimum_snr=3.0,
-        scene_minimum_anchors=2,
-        scene_coupling_thresh=0.01,
-        astrom_kwargs={"poly": {"order": 0}, "gp": {"length_scale": 400}},
-        # synthetic data with a known shift: measure the solver undamped, so the
-        # assertion is on the recovered shift and not on the damping schedule
-        astrom_damping=1.0,
-    )
-
-    scenes = _run_scenes(tmpls.templates, science, weight, cfg)
-    dx_arr, dy_arr = _collect_to_shifts(scenes)
-
-    print(f"\n[test1] True shift: dx={true_dx:.3f}, dy={true_dy:.3f} px")
-    print(f"[test1] Recovered (n={len(dx_arr)}): "
-          f"dx={dx_arr.mean():.3f}±{dx_arr.std():.3f}, "
-          f"dy={dy_arr.mean():.3f}±{dy_arr.std():.3f}")
-
-    assert len(dx_arr) > 0, "No shifts were recovered — check bright-source threshold"
-    assert abs(dx_arr.mean() - true_dx) < 0.3, (
-        f"dx recovered {dx_arr.mean():.3f}, expected {true_dx:.3f}"
-    )
-    assert abs(dy_arr.mean() - true_dy) < 0.3, (
-        f"dy recovered {dy_arr.mean():.3f}, expected {true_dy:.3f}"
-    )
-
-
-# ---------------------------------------------------------------------------
-# Test 2: shift recovery degrades with PSF width
-# ---------------------------------------------------------------------------
-
-def test_scene_shift_uncertainty_scales_with_psf_width():
-    """Shift recovery is more accurate with a narrower PSF (stronger gradient signal).
-
-    The same true shift is injected into two science images: one convolved with a
-    narrow PSF (F770W-like, FWHM≈6 px) and one with a wide PSF (F1800W-like,
-    FWHM≈15 px).  The narrow-PSF case should have a smaller shift-recovery error.
-    """
-    images, segmap, catalog, psfs, truth, wht = make_simple_data(
-        nsrc=20, size=151, peak_snr=50, seed=7
-    )
-    psf_hi = PSF.from_array(psfs[0])
-
-    narrow_arr = _make_gaussian_psf_array(fwhm=6.0, size=51)
-    wide_arr = _make_gaussian_psf_array(fwhm=15.0, size=51)
-
-    science_narrow = fftconvolve(truth, narrow_arr, mode="same")
-    science_wide = fftconvolve(truth, wide_arr, mode="same")
-
-    rng = np.random.default_rng(7)
-    noise_std = float(truth.max()) / 50.0
-    science_narrow += rng.normal(0, noise_std, science_narrow.shape)
-    science_wide += rng.normal(0, noise_std, science_wide.shape)
-    weight = np.ones_like(truth) / noise_std ** 2
-
-    true_dx, true_dy = 0.4, -0.35
-    science_narrow = nd_shift(science_narrow, (true_dy, true_dx))
-    science_wide = nd_shift(science_wide, (true_dy, true_dx))
-
-    cfg = FitConfig(
-        fit_astrometry_joint=True,
-        astrom_minimum_snr=3.0,
-        scene_minimum_anchors=2,
-        scene_coupling_thresh=0.01,
-        astrom_kwargs={"poly": {"order": 0}, "gp": {"length_scale": 400}},
-        # synthetic data with a known shift: measure the solver undamped, so the
-        # assertion is on the recovered shift and not on the damping schedule
-        astrom_damping=1.0,
-    )
-
-    positions = list(zip(catalog["x"], catalog["y"]))
-
-    kernel_narrow = psf_hi.matching_kernel(PSF.from_array(narrow_arr))
-    tmpls_narrow = Templates.from_image(images[0], segmap, positions, kernel_narrow)
-    scenes_narrow = _run_scenes(tmpls_narrow.templates, science_narrow, weight, cfg)
-    dx_n, dy_n = _collect_to_shifts(scenes_narrow)
-
-    kernel_wide = psf_hi.matching_kernel(PSF.from_array(wide_arr))
-    tmpls_wide = Templates.from_image(images[0], segmap, positions, kernel_wide)
-    scenes_wide = _run_scenes(tmpls_wide.templates, science_wide, weight, cfg)
-    dx_w, dy_w = _collect_to_shifts(scenes_wide)
-
-    err_narrow = float(np.sqrt((dx_n.mean() - true_dx) ** 2 + (dy_n.mean() - true_dy) ** 2))
-    err_wide = float(np.sqrt((dx_w.mean() - true_dx) ** 2 + (dy_w.mean() - true_dy) ** 2))
-
-    print(f"\n[test2] True shift: dx={true_dx:.3f}, dy={true_dy:.3f} px")
-    print(f"[test2] Narrow PSF (FWHM=6): dx={dx_n.mean():.3f}, dy={dy_n.mean():.3f} — error={err_narrow:.4f} px")
-    print(f"[test2] Wide PSF  (FWHM=15): dx={dx_w.mean():.3f}, dy={dy_w.mean():.3f} — error={err_wide:.4f} px")
-
-    assert len(dx_n) > 0 and len(dx_w) > 0, "No shifts recovered in one of the PSF cases"
-    assert err_narrow < err_wide, (
-        f"Expected narrow PSF error ({err_narrow:.4f}) < wide PSF error ({err_wide:.4f})"
-    )
-
-
-# ---------------------------------------------------------------------------
-# Test 3: shift estimate scales as 1/alpha0 — biased flux biases shifts
-# ---------------------------------------------------------------------------
 
 def test_scene_shift_depends_on_alpha0_scale():
-    """Shift estimate ∝ 1/alpha0_scale: underestimated fluxes inflate shifts.
+    """Shift estimate is proportional to 1/alpha0: biased fluxes bias shifts.
 
-    The shift signal is bB ∝ alpha0 × image_gradient and the gradient information
-    matrix BB ∝ alpha0².  Their ratio shift ≈ bB/BB ∝ 1/alpha0.  When alpha0 is
-    scaled by factor k, the recovered shift should scale by 1/k.  This confirms
-    that the old astrom_reg flux bias (alpha0 ≈ 0.57×) was inflating shifts by ~1.75×.
+    The shift signal is bB proportional to alpha0 x image_gradient and the
+    gradient information matrix BB to alpha0 squared. Their ratio
+    shift ~ bB/BB goes as 1/alpha0. Scaling alpha0 by k must scale the
+    recovered shift by 1/k -- which is how the old astrom_reg flux bias
+    (alpha0 ~ 0.57x) was inflating shifts by ~1.75x.
     """
+    from mophongo.scene import Scene
+
     images, segmap, catalog, psfs, truth, wht = make_simple_data(
         nsrc=10, size=101, peak_snr=30, seed=55
     )
@@ -366,9 +207,9 @@ def test_scene_shift_depends_on_alpha0_scale():
         astrom_kwargs={"poly": {"order": 0}, "gp": {"length_scale": 400}},
     )
 
-    # Build a single scene from all templates so the has_shift >= 2 requirement is met.
-    # This keeps the test focused on the alpha0 → shift dependency, not scene partitioning.
-    from mophongo.scene import Scene
+    # One scene from all templates, so the has_shift >= 2 requirement is met
+    # and the test stays on the alpha0 -> shift dependency rather than on
+    # scene partitioning.
     all_templates = list(tmpls.templates)
     A, b, _ = build_normal(all_templates, science, weight)
     scene = Scene(
@@ -388,170 +229,26 @@ def test_scene_shift_depends_on_alpha0_scale():
 
     recovered = {}
     for scale in (0.5, 1.0, 2.0):
-        alpha0 = alpha0_true * scale
         AB, BB, bB = assemble_scene_system_AB(
             scene.templates, science, weight, basis,
-            alpha0=alpha0, order=0, include_y=True,
+            alpha0=alpha0_true * scale, order=0, include_y=True,
         )
         sol = SceneFitter.solve(A, b, AB=AB, BB=BB, bB=bB, config=cfg)
         if sol.shifts is not None and len(sol.shifts) >= 2:
             # order=0: shifts = [beta_x, beta_y], each length p=1
             recovered[scale] = (float(sol.shifts[0]), float(sol.shifts[1]))
 
-    print(f"\n[test3] True shift: dx={true_dx:.3f}, dy={true_dy:.3f} px")
-    for scale, (rdx, rdy) in recovered.items():
-        print(f"[test3] alpha0 scale={scale:.1f}: dx={rdx:.4f}, dy={rdy:.4f}")
-
-    assert set(recovered) >= {0.5, 1.0, 2.0}, "Not all alpha0 scales produced shift estimates"
+    assert set(recovered) >= {0.5, 1.0, 2.0}, "not all alpha0 scales produced a shift"
 
     dx_half, _ = recovered[0.5]
     dx_one, _ = recovered[1.0]
     dx_two, _ = recovered[2.0]
 
-    # shift ∝ 1/scale: half alpha0 → double shift, double alpha0 → half shift
-    ratio_half = dx_half / dx_one if abs(dx_one) > 1e-6 else float("nan")
-    ratio_two = dx_two / dx_one if abs(dx_one) > 1e-6 else float("nan")
-    print(f"[test3] dx(0.5×)/dx(1×) = {ratio_half:.3f}  (expected ~2.0)")
-    print(f"[test3] dx(2×) /dx(1×) = {ratio_two:.3f}  (expected ~0.5)")
-    assert abs(ratio_half - 2.0) < 0.5, f"Expected ratio ≈ 2, got {ratio_half:.3f}"
-    assert abs(ratio_two - 0.5) < 0.25, f"Expected ratio ≈ 0.5, got {ratio_two:.3f}"
+    ratio_half = dx_half / dx_one
+    ratio_two = dx_two / dx_one
+    assert abs(ratio_half - 2.0) < 0.5, f"expected ratio ~2, got {ratio_half:.3f}"
+    assert abs(ratio_two - 0.5) < 0.25, f"expected ratio ~0.5, got {ratio_two:.3f}"
 
-    # The unbiased (scale=1.0) estimate should match the true shift
-    assert abs(dx_one - true_dx) < 0.3, f"Unbiased dx={dx_one:.4f}, true={true_dx}"
-    assert abs(recovered[1.0][1] - true_dy) < 0.3, (
-        f"Unbiased dy={recovered[1.0][1]:.4f}, true={true_dy}"
-    )
-
-
-# ---------------------------------------------------------------------------
-# Test 4: iterative shift application converges (residual shift shrinks)
-# ---------------------------------------------------------------------------
-
-def test_scene_shift_iteration_converges():
-    """Residual shifts in the second iteration are smaller than in the first.
-
-    After applying the first-pass shifts to the templates, the second pass should
-    see only small residual offsets.  If the solver is diverging, second-pass
-    |to_shift| would be larger than or equal to the first pass — a red flag for
-    the iterative scheme used by fit_astrometry_niter=2.
-    """
-    images, segmap, catalog, psfs, truth, wht = make_simple_data(
-        nsrc=15, size=151, peak_snr=20, seed=77
-    )
-
-    psf_hi = PSF.from_array(psfs[0])
-    psf_lo = PSF.from_array(psfs[1])
-    kernel = psf_hi.matching_kernel(psf_lo)
-
-    true_dx, true_dy = 0.5, -0.4
-    science = nd_shift(images[1], (true_dy, true_dx))
-    weight = wht[1]
-
-    positions = list(zip(catalog["x"], catalog["y"]))
-    tmpls = Templates.from_image(images[0], segmap, positions, kernel)
-
-    cfg = FitConfig(
-        fit_astrometry_joint=True,
-        astrom_minimum_snr=3.0,
-        scene_minimum_anchors=2,
-        scene_coupling_thresh=0.01,
-        astrom_kwargs={"poly": {"order": 0}, "gp": {"length_scale": 400}},
-        # synthetic data with a known shift: measure the solver undamped, so the
-        # assertion is on the recovered shift and not on the damping schedule
-        astrom_damping=1.0,
-    )
-
-    # --- Pass 1: solve, record to_shift, then apply ---
-    scenes1 = _run_scenes(tmpls.templates, science, weight, cfg)
-    dx1, dy1 = _collect_to_shifts(scenes1)
-    mag1 = float(np.sqrt(dx1 ** 2 + dy1 ** 2).mean()) if len(dx1) > 0 else 0.0
-
-    # Apply the first-pass shifts to the templates in-place
-    all_templates = [t for s in scenes1 for t in s.templates]
-    Templates.apply_template_shifts(all_templates)
-
-    # --- Pass 2: rebuild normal equations from shifted templates, solve again ---
-    scenes2 = _run_scenes(tmpls.templates, science, weight, cfg)
-    dx2, dy2 = _collect_to_shifts(scenes2)
-    mag2 = float(np.sqrt(dx2 ** 2 + dy2 ** 2).mean()) if len(dx2) > 0 else 0.0
-
-    print(f"\n[test4] True shift: dx={true_dx:.3f}, dy={true_dy:.3f} px")
-    print(f"[test4] Pass 1: mean |shift| = {mag1:.4f} px  "
-          f"(dx={dx1.mean():.3f}, dy={dy1.mean():.3f})")
-    print(f"[test4] Pass 2: mean |shift| = {mag2:.4f} px  "
-          f"(dx={dx2.mean():.3f}, dy={dy2.mean():.3f})")
-
-    assert len(dx1) > 0, "Pass 1 produced no shifts"
-    assert mag1 > 0.05, f"Pass-1 shift too small ({mag1:.4f}), test may not be exercising the solver"
-    assert mag2 < mag1, (
-        f"Pass-2 residual shift ({mag2:.4f} px) should be smaller than pass-1 ({mag1:.4f} px)"
-    )
-
-
-def test_leverage_cap_bounds_one_bright_anchor(tmp_path):
-    """A bright lopsided source must not carry the scene's shift on its own.
-
-    Anchor leverage in the shift block goes as flux squared, so one bright
-    source outweighs many faint ones. When that source is extended with an
-    asymmetric colour gradient its residual is a dipole aligned with its own
-    template gradient -- formally a shift -- and it drags the fitted field.
-    Capping leverage bounds its influence while leaving the shift it measures
-    untouched.
-    """
-    import numpy as np
-    import pytest
-    from mophongo.psf import PSF
-    from mophongo.scene import assemble_scene_system_AB, make_scene_basis
-    from mophongo.templates import Templates
-
-    size, npsf = 161, 31
-    rng = np.random.default_rng(5)
-    psf = PSF.gaussian(npsf, 3.0).array
-    # one bright source plus eight faint ones on a ring around it
-    pos = [(80.0, 80.0)] + [
-        (80.0 + 45 * np.cos(t), 80.0 + 45 * np.sin(t))
-        for t in np.linspace(0, 2 * np.pi, 8, endpoint=False)
-    ]
-    flux = np.array([300.0] + [3.0] * 8)
-
-    hires = np.zeros((size, size))
-    for (x, y), f in zip(pos, flux):
-        sly = slice(int(y) - npsf // 2, int(y) + npsf // 2 + 1)
-        slx = slice(int(x) - npsf // 2, int(x) + npsf // 2 + 1)
-        hires[sly, slx] += f * psf
-    yy, xx = np.indices(hires.shape)
-    near = np.argmin([np.hypot(xx - x, yy - y) for x, y in pos], axis=0)
-    segmap = np.where(hires > 1e-3 * hires.max(), near + 1, 0).astype(np.int32)
-    tmpls = Templates.from_image(hires, segmap, pos, kernel=PSF.gaussian(npsf, 1.0).array)
-    templates = list(tmpls)
-
-    # image: every source truly unshifted, except the bright one, which is
-    # given a 1-pixel lopsidedness standing in for a colour gradient
-    image = np.zeros_like(hires)
-    for t, f in zip(templates, flux):
-        image[t.slices_original] += f * t.data[t.slices_cutout]
-    t0 = templates[0]
-    gy0, gx0 = np.gradient(t0.data.astype(float))
-    image[t0.slices_original] += -1.0 * flux[0] * gx0[t0.slices_cutout]
-    weights = np.ones_like(image)
-    image += rng.normal(0, 1e-6, image.shape)
-
-    alpha0 = np.asarray(flux, dtype=float)
-    is_bright = np.ones(len(templates), dtype=bool)
-    basis, _c, _s = make_scene_basis(templates, is_bright, order=0)
-
-    def solve(cap):
-        _AB, BB, bB = assemble_scene_system_AB(
-            templates, image, weights, basis, alpha0=alpha0, order=0,
-            include_y=True, leverage_cap=cap,
-        )
-        BB = BB.toarray()
-        return float(np.linalg.solve(BB + 1e-12 * np.eye(BB.shape[0]), bB)[0])
-
-    dx_free = solve(None)
-    dx_capped = solve(0.5)          # clip anything above the median anchor
-    # the bright source pulls the whole scene; the cap pulls it back toward 0
-    assert abs(dx_free) > 0.5
-    assert abs(dx_capped) < 0.5 * abs(dx_free)
-    # and the cap is inert when every anchor is already below it
-    assert solve(1.0) == pytest.approx(dx_free, rel=1e-12)
+    # the unbiased (scale=1.0) estimate matches the injected shift
+    assert abs(dx_one - true_dx) < 0.3
+    assert abs(recovered[1.0][1] - true_dy) < 0.3
