--- tests/test_fit.py
@@ -1,30 +1,0 @@
-import os
-import sys
-
-sys.path.insert(0, os.path.join(os.path.dirname(__file__), "..", "src"))
-
-import numpy as np
-
-from mophongo.fit import FitConfig, SparseFitter
-from mophongo.templates import Templates
-from utils import make_simple_data
-
-
-def test_flux_and_rms_estimation():
-    """SparseFitter.flux_and_rms matches quick flux and error estimates."""
-    images, segmap, catalog, psfs, _, rms = make_simple_data()
-
-    tmpls = Templates.from_image(
-        images[0], segmap, list(zip(catalog["x"], catalog["y"])), kernel=None
-    )
-
-    fitter = SparseFitter(tmpls.templates, images[1], 1.0 / rms[1] ** 2, FitConfig())
-
-    flux, err = fitter.flux_and_rms()
-    np.testing.assert_allclose(flux, fitter.quick_flux())
-    np.testing.assert_allclose(err, fitter.predicted_errors())
-
-    for tmpl in tmpls.templates:
-        tmpl.flux = 42.0
-    flux2, _ = fitter.flux_and_rms()
-    assert np.all(flux2 == 42.0)
